+++ test/unit/network/conftest.py
@@ -1,0 +1,10 @@
+# coding: utf-8
+
+from __future__ import unicode_literals, absolute_import
+
+import pytest
+
+
+@pytest.fixture(params=('GET', 'POST', 'PUT', 'DELETE', 'OPTIONS'))
+def http_verb(request):
+    return request.param